Removed logs, and rearranged some functions and separated components.
