+++ client/src/lib/api.js
@@ -15,3 +15,21 @@ export default async function getMonthsRecords (tokenKey) {
   console.error(err);
  }
 }
+
+/**
+ * Sends the server request to sign up a new user.
+ * @param {object} user object with all fields for a new user.
+ * @returns the response from the server
+ */
+export async function sendSignUp(user) {
+  const response = await fetch('/api/home/sign-up', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(user)
+  });
+  if (!response.ok) throw new Error(`Invalid New User.`, { cause: response })
+  const confirm = await response.json();
+  return confirm;
+}
